Fixed bug where the display would fail to render completely after doing Rendering->Update on large graphs.  This could be reproduced by choosing Rendering->Update twice in quick succession with a large graph loaded---the second update should be timed to occur before the rendering caused by the first update is finished.
The problem was that H3Transformer was using space in H3Graph to store
graph traversal state.  This was fine when Walrus had only one instance of
H3Transformer for the duration of a program run, but this no longer is the
case, and storing the traversal state in H3Graph causes old data to persist
and disrupt new instances of H3Transformer.  The solution was to move this
state into H3Transformer.