Made the rotation state interruptible.  That is, an external thread can cause the state to change in H3AdaptiveRenderLoop while the latter is in STATE_ROTATE.  Previously, an H3AdaptiveRenderLoop remained in STATE_ROTATE until the invoking H3RotationRequest finished.
This change was made so that a thread could abort an ongoing rotation by
simply invoking an appropriate method in the H3RenderLoop interface.  However,
because of the wide ramifications of this change, I've decided against this
approach to aborting rotations.  There are outright problems with it right
now; for example, the display can lockup if the user starts a new rotation
quickly after stopping a previous rotation.  This is caused by
H3RenderLoop.rotateDisplay() being run while H3AdaptiveRenderLoop is still
in STATE_ROTATE.  There are other problems elsewhere.

At any rate, this code may prove useful in the future, so it's being preserved
in this branch.