Added a new top-level menu, "Display", which contains menu items for interacting with Walrus in various ways when a graph is being rendered. Hooked up the following menu items: Refresh, Wobble, Save Position, and Restore Position.  The other menu items are enabled and disabled properly, though not actually hooked up to any actions.
As part of implementing the above menu items, I added forceIdleState() to
EventHandler.  This method causes EventHandler to cleanly stop whatever it is
doing (such as rotations or wobbling) and return to the idle state.  This
is needed in various places since rotations aren't interruptible in
implementations of H3RenderLoop.  Without this method, Walrus would lock
up if the user chose, say, "Stop" from the Rendering menu while the display
was wobbling.